Create actual .md files and implement usage
