+++ app/components/terminal/bootSequence.ts
