+++ app/components/terminal/commands.ts
