+++ app/components/terminal/types.ts
