+++ app/projects/portfolio.md
@@ -1,0 +1,28 @@
+# Portfolio
+
+## Description
+An interactive terminal-style portfolio (you are here!).
+
+## Technologies Used
+- React
+- Next.js
+- TypeScript
+- Tailwind CSS
+
+## Features
+- Simulated file system
+- Custom command processing
+- Markdown rendering
+- Responsive design
+
+## Challenges Overcome
+- Implementing a virtual file system
+- Creating an authentic terminal experience
+- Balancing authenticity with user-friendliness
+
+## Future Improvements
+- Add more interactive elements
+- Implement a simple game or Easter egg
+- Create a "GUI" version toggle
+
+[View Project](https://github.com/zukauskas/portfolio)`
