+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "generate-project-data": "node app/scripts/generateProjectData.js"
   },
   "dependencies": {
     "@radix-ui/react-icons": "^1.3.0",
